pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cisco-ai-a2a-scanner"
version = "1.0.0"
description = "A comprehensive security scanner for Agent-to-Agent (A2A) protocol threats"
authors = [
    {name = "Cisco"},
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Security",
]
dependencies = [
    "httpx>=0.28.1",
    "yara-python>=4.5.4",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "litellm>=1.77.0",
    "aiohttp>=3.9.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "jsonschema>=4.20.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "respx>=0.22.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
a2a-scanner = "a2ascanner.cli:main"
a2a-scanner-api = "a2ascanner.api.server:main"
a2ascanner-proxy = "a2ascanner.proxy.cli:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "--maxfail=1 -v"
asyncio_mode = "auto"
filterwarnings = [
    # Silence Pydantic serializer warnings emitted by LiteLLM models
    "ignore:Pydantic serializer warnings:UserWarning:pydantic.*",
]
markers = [
    "integration: mark for tests that hit external services (run with -m integration)",
]

[tool.setuptools.packages.find]
include = ["a2ascanner*", "examples*", "data*"]

[tool.setuptools.package-data]
"a2ascanner" = ["data/prompts/*.md", "data/yara_rules/*.yara", "data/yara_rules/*.yar"]

# Code formatting and linting configuration
[tool.black]
line-length = 88
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503", "E501"]
exclude = [".git", "__pycache__", "build", "dist", ".venv", "venv"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "yara",
    "litellm",
]
ignore_missing_imports = true
